backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv